+++ src/main/java/org/jolokia/docker/maven/util/EnvUtil.java
@@ -170,11 +170,19 @@ public class EnvUtil {
         return System.getenv("DOCKER_REGISTRY");
     }
 
-    public static File prepareAbsolutePath(MojoParameters params, String path) {
+    public static File prepareAbsoluteOutputDirPath(MojoParameters params, String dir, String path) {
+        return prepareAbsolutePath(params, new File(params.getOutputDirectory(), dir).toString(), path);
+    }
+    
+    public static File prepareAbsoluteSourceDirPath(MojoParameters params, String path) {
+        return prepareAbsolutePath(params, params.getSourceDirectory(), path);
+    }
+        
+    private static File prepareAbsolutePath(MojoParameters params, String directory, String path) {
         File file = new File(path);
         if (file.isAbsolute()) {
             return file;
         }
-        return new File(new File(params.getProject().getBasedir(), params.getSourceDirectory()), path);
+        return new File(new File(params.getProject().getBasedir(), directory), path);
     }
 }