+++ doc/changelog.md
@@ -4,6 +4,7 @@
   - Fix maven parse error when specifying restart policy (#99)
   - Allow hostnames to be used in port bindings (#101)
   - Add support for tagging at build and push time (#104)
+  - Use correct output dir during multi-project builds (#97)
 
 * **0.11.1**
   - Add support for binding UDP ports (#83)